requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
discord.py @ git+https://github.com/Rapptz/discord.py@799e3c5bf09cedcd9b7e2b83df733910fa11532a
dnspython==2.2.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
pymongo==4.3.3
yarl==1.8.2
